Re-enable staticcheck: fix broken config and all surfaced findings
staticcheck.conf's checks list lacked "inherit", which in staticcheck's
config semantics REPLACES the default check set — so CI's staticcheck
step has reported nothing since the file was added. Fix the config
(inherit, -S1008, -ST1005) and everything the working checker surfaces:

- Delete dead code (U1000): leftover copies of the gather-metadata
  functions in assessMigrationCommand.go (moved to src/migassessment in
  an earlier PR), the unused config-file key readers in config.go, and
  a dozen other unused funcs/vars/fields.
- Split cmd/importDataTestUtils.go: shared batch-producer helpers keep
  the file (now tagged unit || integration || integration_voyager_command);
  the YugabyteDB-container helpers move to
  importDataIntegrationTestUtils.go (integration tags only). Previously
  the file was untagged, so its integration-only helpers looked dead to
  any single-tag analysis.
- Check errors in importDataFileTaskPicker_test.go that were assigned
  and silently overwritten (SA4006).
- Mechanical simplifications: redundant returns, merged declarations,
  for-range over channel/keys, type-switch bindings, unwrapped nil
  checks (S1000/S1005/S1011/S1021/S1023/S1028/S1031/S1034/S1039).
- CI (go.yml): also run vet and staticcheck with -tags unit — all 110
  test files are tag-gated, so the untagged run never analyzed them.
- Fix //go:build unit_test -> unit on src/utils/type_mapping_test.go;
  that test file was invisible to CI's -tags unit test run.
- Delete stray .github/workflows/Untitled; document the lint commands
  and the "inherit" pitfall in AGENTS.md.
