+++ yb-voyager/cmd/importDataFileTaskPicker_test.go
@@ -195,6 +195,7 @@ func TestSequentialTaskPickerResumePicksInProgressTask(t *testing.T) {
 	// simulate restart by creating a new picker
 	slices.Reverse(tasks) // reorder the tasks so that the in progress task is at the end
 	picker, err = NewSequentialTaskPicker(tasks, state)
+	assert.NoError(t, err)
 
 	// no matter how many times we call NextTask, it should return the same task (first task)
 	for i := 0; i < 10; i++ {
@@ -1190,6 +1191,7 @@ func TestColocatedAwareRandomTaskPickerResumable(t *testing.T) {
 	task1, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp1, err := NewSequentialFileBatchProducer(task1, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch1, err := fbp1.NextBatch()
 	assert.NoError(t, err)
 	batch1.MarkInProgress()
@@ -1197,6 +1199,7 @@ func TestColocatedAwareRandomTaskPickerResumable(t *testing.T) {
 	task2, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp2, err := NewSequentialFileBatchProducer(task2, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch2, err := fbp2.NextBatch()
 	assert.NoError(t, err)
 	batch2.MarkInProgress()
@@ -1204,6 +1207,7 @@ func TestColocatedAwareRandomTaskPickerResumable(t *testing.T) {
 	task3, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp3, err := NewSequentialFileBatchProducer(task3, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch3, err := fbp3.NextBatch()
 	assert.NoError(t, err)
 	batch3.MarkInProgress()
@@ -1721,7 +1725,7 @@ func TestColocatedCappedRandomTaskPickerMultipleTasksColocatedAndSharded(t *test
 	}
 	// empty queue
 	for i := 0; i < 4; i++ {
-		_ = <-colocatedBatchImportQueue
+		<-colocatedBatchImportQueue
 	}
 	// now only one new colocated task should be picked.
 	newColocatedTask, err := picker.Pick()
@@ -2058,6 +2062,7 @@ func TestColocatedCappedRandomTaskPickeResumable(t *testing.T) {
 	task1, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp1, err := NewSequentialFileBatchProducer(task1, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch1, err := fbp1.NextBatch()
 	assert.NoError(t, err)
 	batch1.MarkInProgress()
@@ -2065,6 +2070,7 @@ func TestColocatedCappedRandomTaskPickeResumable(t *testing.T) {
 	task2, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp2, err := NewSequentialFileBatchProducer(task2, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch2, err := fbp2.NextBatch()
 	assert.NoError(t, err)
 	batch2.MarkInProgress()
@@ -2072,6 +2078,7 @@ func TestColocatedCappedRandomTaskPickeResumable(t *testing.T) {
 	task3, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp3, err := NewSequentialFileBatchProducer(task3, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch3, err := fbp3.NextBatch()
 	assert.NoError(t, err)
 	batch3.MarkInProgress()
@@ -2079,6 +2086,7 @@ func TestColocatedCappedRandomTaskPickeResumable(t *testing.T) {
 	task4, err := picker.Pick()
 	assert.NoError(t, err)
 	fbp4, err := NewSequentialFileBatchProducer(task4, state, false, errorHandler, progressReporter)
+	assert.NoError(t, err)
 	batch4, err := fbp4.NextBatch()
 	assert.NoError(t, err)
 	batch4.MarkInProgress()
