+++ yb-voyager/src/query/queryissue/detectors.go
@@ -719,7 +719,7 @@ func (d *DatabaseOptionsDetector) Detect(msg protoreflect.Message) error {
 	if err != nil {
 		return err
 	}
-	for defName, _ := range defNames {
+	for defName := range defNames {
 		if unsupportedDatabaseOptionsFromPG15.ContainsOne(defName) {
 			d.pg15OptionsDetected.Add(defName)
 		}
