+++ yb-voyager/src/utils/type_mapping_test.go
@@ -1,4 +1,4 @@
-//go:build unit_test
+//go:build unit
 
 package utils
