+++ yb-voyager/staticcheck.conf
@@ -1,2 +1,6 @@
-# Disable the S1008 check to allow more explicit boolean returns
-checks = ["-S1008"]
+# "inherit" keeps staticcheck's default check set; without it, the list below
+# REPLACES the defaults and every check is silently disabled.
+# S1008:  disabled to allow more explicit boolean returns.
+# ST1005: disabled; error strings in this CLI are shown directly to users and
+#         are deliberately capitalized/punctuated.
+checks = ["inherit", "-S1008", "-ST1005"]
